Use a State enum to keep track of the resource state
Using a State enum to keep track of the resource state instead of
relying on the nullability of result and error allows the result and
error to be null. It is totally possible that a resource may be some
object or null, so we should allow it to be null and not rely on null as
an indicator that we don't have the result yet. Same goes for the error.
Albeit rare, it is totally possible for the promise rejection error to
be null, so we should allow for that.